Parsing out the Due Date
Now when you add a task, you can add a due date to it. Its not 100
percent yet, right now the Regex is not fully set up so currently for
the due date it will grab the current year, Ive hardcoded the month to
be May, but the day is the only one set up correctly right now. Next
Steps: To add on to the regex to grab whatever year the user inputs as
well as any proper month they put in. The more I watch these videos the
more I learn and the more I want to create on my own!!